Added music
I got it to loop and I did it the scripting way, I think... Check out Earl scene 4. Also check disc for some ideas on SFX.
