+++ Assets/Scripts/UX/Sound/Music.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 14307c1a5d26c5f44a0a485dc8c886b7
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
