+++ Assets/Scripts/UX/Sound/Music/MusicManager.cs
@@ -1,0 +1,15 @@
+using UnityEngine;
+
+public class MusicManager : MonoBehaviour
+{
+    public AudioClip MainMusic;
+    private AudioSource audioSource;
+
+    void Start()
+    {
+        audioSource = GetComponent<AudioSource>();
+        audioSource.clip = MainMusic;
+        audioSource.loop = true;
+        audioSource.Play();
+    }
+}
