+++ Assets/Scripts/UX/Sound/SFX.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: eb34507c67792984981dbb6eb74899b0
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
