+++ Assets/Scripts/UX/Sound/SFX/SoundFXManager.cs
@@ -1,0 +1,30 @@
+using Unity.Mathematics;
+using UnityEngine;
+
+public class SoundFXManager : MonoBehaviour
+{
+    [SerializeField] private AudioSource soundFXObject;
+    public static SoundFXManager instance;
+
+    private void Awake() {
+        if (instance == null) {
+            instance = this;
+        }
+    }
+
+    public void PlaySoundFXClip(AudioClip audioClip, Transform spawnTransform, float volume) {
+        AudioSource audioSource = Instantiate(soundFXObject, spawnTransform.position, quaternion.identity);
+
+        audioSource.clip = audioClip;
+
+        audioSource.volume = volume;
+
+        audioSource.Play();
+
+        float clipLength = audioSource.clip.length;
+
+        Destroy(audioSource.gameObject, clipLength);
+
+
+    }
+}
