--- Assets/Scripts/UX/Sound/SoundFXManager.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: fa277d827478d3149842f74f979d1e8d
